refactor: refactor prj structure
