--- lqm/week2/goroutine/worker/job.go
@@ -1,11 +1,0 @@
-package worker
-
-import "context"
-
-type JobHandler func(ctx context.Context) Result
-
-type Job struct {
-	ID      int
-	Payload string
-	Handler JobHandler
-}
