--- lqm/week2/goroutine/worker/result.go
@@ -1,6 +1,0 @@
-package worker
-
-type Result struct {
-	JobID int
-	State int // 0: Failed, 1: Success
-}
